[BC-19] Add MSS provider account configuration
Description
-----------

This commit adds MSS provider account configuration to
Tergite-Qiskit-connector. Previously, the MSS URL was hardcoded.

The provider account configuration is stored at file
  > $HOME/.qiskit/tergiterc
under section(s) [service <service name>].
It is possible to provide a configuration for multiple services.

Example:
--------

  > [service mss 1]
  > url = http://qdp-git.mc2.chalmers.se:5000
  > token = None
  > verify = False
  >
  > [service mss 2]
  > url = http://qdp-git.mc2.chalmers.se:5001
  > note = Secondary service on port 5001

Currently, only the field 'url' is obligatory. Other fields
are parsed and stored, but not used in otherwise.

A specific provider account can be selected via:
  > provider = Tergite.get_provider('service name')
Service name example is 'mss 1' as seen in the above tergiterc.

Shall no service name be given, eg:
  > provider = Tergite.get_provider()
The first service found in tergiterc will be used.

Create new provider account configuration
-----------------------------------------

Shall the tergiterc file not provide any valid service section
a warning will be printed. A new provider account configuration can
be created on-fly with the following code:

  > from tergite.providers.tergite.provider_account import ProviderAccount
  > my_account = ProviderAccount("service name", "http://myaddress.com:port")
  > provider = Tergite.use_provider_account(my_account)

For later convenience, it is possibele to store the new account in the
tergiterc file with "safe=True" option.

  > provider = Tergite.use_provider_account(my_account, save=True)

Verification
------------
This functionality has been verified:
1) The file tergiterc does not exists, create and use a new provider
   account on fly
2) The same as above, plus save the account for later convenience.
3) Only one service defined in tergiterc, it is picked-up by
   Tergite.get_provider()
4) Multiple services are defined in tergiterc, the first one is
   picked-up by Tergite.get_provider()
5) Multiple services are defined in tergiterc, a named one is
   selected by Tergite.get_provider("service name")
6) From a provider, backend is selected via
   provider.get_backend("name"). The backend has properly set .url
   attribute.
7) Functionality with the whole stack: Qiskit, MSS, BCC, Labber has been
   verified.