resotred the "/demo" and "/dev" dirc. to original
The "demo" and "/dev" directory is resored to its
last original commit 18ff53083fd94943b2adfe5579ef7cc84f6374fc,
which is the version uses "Calibraion Tables" for backend
properties fetch, which is conducted by several http
requests. This is a complicated process of involving
calibration jobs, calibraion types and so on.
In this version the backend properties are read from
mongo DB in a simpler way. A "test_job_example" folder
is created to test the jobs through the SW stack. These
example jobs are tested with stubbed version of MSS and
BCC that moks the interaction between processes in the
SW stack.