Fix IndexError when coupling_dict has qubits whose ids don't start at 0
Imagine you have your qubits: qubit_ids = ["q17", "q18", "q19"]

Originally, the coupling map would just take "q17" to be 0, "q18" to be 1, "q19" to be 2.
Thus the coupling_map would be [[0, 1], [1, 0], ...]
However, with the coupling_dict, the generated coupling_map is [[17, 18], [18, 19],...]

When this config is passed to tergite SDK, it runs into IndexErrors
because [17,18] does not exist in the qubits because
the call to "calibrations.add_instructions" passes
qubits=tuple(q for q in range(self.data["num_qubits"])), which completely ignores the "17", "18" etc.